fix(core): keep cache helper out of ACP bundle
